Add builder module and robot builder
