+++ config/config.yaml
@@ -23,17 +23,15 @@ plugins:
     dispatcher:
       re-allocate: True
 
-robot:
-  components:
-    stp_solver: srea
-    bidder:
-      bidding_rule:
-        robustness: srea  # has to be the same as the stp_solver
-        temporal: completion_time
-      auctioneer_name: fms_zyre_api # This is completely Zyre dependent
-    schedule_monitor:
-      corrective_measure: re-allocate
-
+robot_proxy:
+  stp_solver: srea
+  bidder:
+    bidding_rule:
+      robustness: srea  # has to be the same as the stp_solver
+      temporal: completion_time
+    auctioneer_name: fms_zyre_api # This is completely Zyre dependent
+  schedule_monitor:
+    corrective_measure: re-allocate
   robot_store:
     db_name: robot_store
     port: 27017
