+++ mrs/task_allocation/bidder.py
@@ -177,15 +177,3 @@ class Bidder(RobotBase):
         self.logger.debug("Robot %s sends close round msg ", self.id)
         self.api.publish(msg, groups=['TASK-ALLOCATION'])
 
-
-class BidderBuilder:
-    def __init__(self):
-        self._instance = None
-
-    def __call__(self, bidder_config, **kwargs):
-        if not self._instance:
-            self._instance = Bidder(bidder_config, **kwargs)
-        return self._instance
-
-
-configure = BidderBuilder()
