+++ setup.py
@@ -3,8 +3,9 @@
 from setuptools import setup
 
 setup(name='mrs',
-      packages=['mrs', 'mrs.config', 'mrs.db.models', 'mrs.db.models.performance', 'mrs.db.queries', 'mrs.structs',
-                'mrs.utils', 'mrs.exceptions', 'mrs.task_allocation', 'mrs.task_execution'],
+      packages=['mrs', 'mrs.config', 'mrs.config.builders', 'mrs.db.models', 'mrs.db.models.performance',
+                'mrs.db.queries', 'mrs.structs', 'mrs.utils', 'mrs.exceptions', 'mrs.task_allocation',
+                'mrs.task_execution'],
       version='0.2.0',
       install_requires=[
             'numpy'
